fix(AvatarStack): displaying more than 9 remaining avatars
