+++ src/components/AvatarStack/types.ts
@@ -41,3 +41,19 @@ export interface AvatarStackProps {
      */
     renderMore?: (options: {count: number}) => React.ReactElement;
 }
+
+export interface AvatarStackMoreProps {
+    className?: string;
+    count: number;
+    'aria-label'?: string;
+    size?: AvatarSize;
+    borderColor?: string;
+}
+
+export type AvatarStackMoreButtonProps = Pick<
+    React.HTMLProps<HTMLButtonElement>,
+    'className' | 'onClick' | 'aria-label'
+> & {
+    size?: AvatarSize;
+    count: number;
+};
